requirements.txt
openai
python-dotenv
pyyaml
scikit-learn
matplotlib
networkx
python-docx
markdown
beautifulsoup4